+++ RealmSwift/ThreadSafeReference.swift
@@ -30,7 +30,7 @@ import Realm
  classes which attempt to conform to it will not make them work with `ThreadSafeReference`.
  */
 public protocol ThreadConfined {
-    // Must also conform to `AssistedObjectiveCBridgeable`
+    // Must also conform to `CustomObjectiveCBridgeable`
 
     /**
      The Realm which manages the object, or `nil` if the object is unmanaged.
@@ -92,8 +92,6 @@ public protocol ThreadConfined {
  - see: `Realm.resolve(_:)`
  */
 @frozen public struct ThreadSafeReference<Confined: ThreadConfined> {
-    private let swiftMetadata: Any?
-
     /**
      Indicates if the reference can no longer be resolved because an attempt to resolve it has
      already occurred. References can only be resolved once.
@@ -111,14 +109,85 @@ public protocol ThreadConfined {
              constructor.
      */
     public init(to threadConfined: Confined) {
-        let bridged = (threadConfined as! AssistedObjectiveCBridgeable).bridged
-        swiftMetadata = bridged.metadata
-        objectiveCReference = RLMThreadSafeReference(threadConfined: bridged.objectiveCValue as! RLMThreadConfined)
+        objectiveCReference = RLMThreadSafeReference(threadConfined: (threadConfined as! CustomObjectiveCBridgeable).objCValue as! RLMThreadConfined)
     }
 
     internal func resolve(in realm: Realm) -> Confined? {
         guard let objectiveCValue = realm.rlmRealm.__resolve(objectiveCReference) else { return nil }
-        return ((Confined.self as! AssistedObjectiveCBridgeable.Type).bridging(from: objectiveCValue, with: swiftMetadata) as! Confined)
+        return ((Confined.self as! CustomObjectiveCBridgeable.Type).bridging(objCValue: objectiveCValue) as! Confined)
+    }
+}
+
+// MARK: ThreadSafe propertyWrapper
+
+/**
+    A property wrapper type that may be passed between threads.
+
+    A `@ThreadSafe` property contains a thread-safe reference to the underlying wrapped value.
+    This reference is resolved to the thread on which the wrapped value is accessed. A new thread
+    safe reference is created each time the property is accessed.
+
+ - warning: This property wrapper should not be used for properties on long lived objects.
+            `@ThreadSafe` properties contain a `ThreadSafeReference` which
+            can pin the source version of the Realm in use. This means that this property
+            wrapper is **better suited for function arguments and local variables**
+            **that get captured by an aynchronously dispatched block.**
+
+ - see: `ThreadSafeReference`
+ - see: `ThreadConfined`
+*/
+@propertyWrapper public final class ThreadSafe<T: ThreadConfined> {
+    private var threadSafeReference: ThreadSafeReference<T>?
+    private var rlmConfiguration: RLMRealmConfiguration?
+    private let lock = NSLock()
+
+    /// :nodoc:
+    public var wrappedValue: T? {
+        get {
+            lock.lock()
+            guard let threadSafeReference = threadSafeReference,
+                  let rlmConfig = rlmConfiguration else {
+                lock.unlock()
+                return nil
+            }
+            do {
+                let rlmRealm = try RLMRealm(configuration: rlmConfig)
+                let realm = Realm(rlmRealm)
+                guard let value = threadSafeReference.resolve(in: realm) else {
+                    self.threadSafeReference = nil
+                    lock.unlock()
+                    return nil
+                }
+                self.threadSafeReference = ThreadSafeReference(to: value)
+                lock.unlock()
+                return value
+            // FIXME: wrappedValue should throw
+            // As of Swift 5.5 property wrappers can't have throwing accessors.
+            } catch let error as NSError {
+                lock.unlock()
+                throwRealmException(error.localizedDescription)
+            }
+        }
+        set {
+            lock.lock()
+            guard let newValue = newValue else {
+                threadSafeReference = nil
+                lock.unlock()
+                return
+            }
+            guard let rlmConfiguration = newValue.realm?.rlmRealm.configuration else {
+                lock.unlock()
+                throwRealmException("Only managed objects may be wrapped as thread safe.")
+            }
+            self.rlmConfiguration = rlmConfiguration
+            threadSafeReference = ThreadSafeReference(to: newValue)
+            lock.unlock()
+        }
+    }
+
+    /// :nodoc:
+    public init(wrappedValue: T?) {
+        self.wrappedValue = wrappedValue
     }
 }
 
@@ -148,3 +217,12 @@ extension Realm {
         return reference.resolve(in: self)
     }
 }
+
+#if swift(>=5.5) && canImport(_Concurrency)
+extension ThreadSafeReference: Sendable {
+}
+extension RLMThreadSafeReference: @unchecked Sendable {
+}
+extension ThreadSafe: @unchecked Sendable {
+}
+#endif